Time: 5 ms (52.92%), Space: 44.3 MB (21.65%) - LeetHub

+++ 0940-distinct-subsequences-ii/0940-distinct-subsequences-ii.java
@@ -10,7 +10,7 @@ class Solution {
         }
         if(dp[n] != -1) return dp[n];
 
-        int total = (2*solve(n-1)%M);
+        int total = (int)(2L*solve(n-1)%M);
         //remove duplicates
         if(prev[n] != 0){
             int duplicates = solve(prev[n] - 1);
